Persist window state with process store
