+++ package.json
@@ -1,0 +1,12 @@
+{
+  "name": "portfolio-os",
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "test": "vitest"
+  },
+  "devDependencies": {
+    "jsdom": "^27.0.0",
+    "vitest": "^1.6.0"
+  }
+}
