+++ vitest.config.js
@@ -1,0 +1,5 @@
+export default {
+  test: {
+    environment: 'jsdom'
+  }
+};
